# CUDA: use async data loading for FlashAttention