--- a/ggml/src/ggml-cuda/mma.cuh
+++ b/ggml/src/ggml-cuda/mma.cuh
@@ -7,7 +7,7 @@
 // A is a row-major matrix with shape M x K.
 // B is a column-major matrix with shape K x N.
 // C is a column-major matrix with shape M x N.
-// A, B, and C are represented using the same fundametal data type: a row-major matrix with I rows and J columns.
+// A, B, and C are represented using the same fundamental data type: a row-major matrix with I rows and J columns.
 // Note that J is measured in physical 32 bit elements instead of logical elements.
 // The methods get_i and get_j can be used to get the physical 32 bit index of the lth element of a thread within a tile.
 // All matrix tiles have ne physical 32 bit elements per warp.