This PR adds the use of asynchronous data copies to the CUDA FlashAttention kernels which enables simultaneous data loading and computation. This feature needs Ampere or newer. I exposed the related PTX instructions via a simple API in a new header `cp-async.cuh` because I don't like the existing CUDA API; I think it obfuscates what the hardware is actually doing and it needlessly causes compilation failures when included with the wrong compute capabilities. While I was working on this I refactored the API in `mma.cuh` to use the same fundamental data structures for all matrix tiles. This is more convenient for kernels where multiple matrix multiplications are being chained and the output tiles of one operation are the input tiles of another operation.

This PR also fixes a bug on master where the condition for when the stream-k fixup for FlashAttention can be skipped was wrong. In practice it is very unlikely though that this bug resulted in incorrect results due to the specific number of streaming multiprocessors found on contemporary NVIDIA GPUs.

<details>
<summary>Performance changes</summary>

| GPU        | Model           |     Microbatch size | Test      |     t/s master |     t/s cuda-fa-mma-17 |     Speedup |
| :--------- | :-------------- | ------------------: | :-------- | -------------: | ---------------------: | ----------: |
| RTX 3090   | gemma 2B F16    |                   8 | pp16384   |         884.99 |                 960.45 |        1.09 |
| RTX 3090   | gemma 2B F16    |                  16 | pp16384   |        1723.90 |                1751.24 |        1.02 |
| RTX 3090   | gemma 2B F16    |                  32 | pp16384   |        2930.83 |                2833.06 |        0.97 |
| RTX 3090   | gemma 2B F16    |                  64 | pp16384   |        3919.85 |                3783.78 |        0.97 |
| RTX 3090   | gemma 2B F16    |                 128 | pp16384   |        7502.87 |                7534.81 |        1.00 |
| RTX 3090   | gemma 2B F16    |                 256 | pp16384   |        9887.99 |                9922.51 |        1.00 |
| RTX 3090   | gemma 2B F16    |                 512 | pp16384   |       11016.88 |               11048.05 |        1.00 |
| RTX 3090   | gemma 2B F16    |                1024 | pp16384   |       11479.26 |               11647.72 |        1.01 |
| RTX 3090   | gemma 2B F16    |                2048 | pp16384   |       11766.60 |               11825.53 |        1.01 |
| RTX 3090   | gemma 2B F16    |                4096 | pp16384   |       11775.10 |               11840.24 |        1.01 |
| RTX 3090   | gemma 2B F16    |                8192 | pp16384   |       11770.49 |               11834.37 |        1.01 |
| RTX 3090   | gemma 2B F16    |               16384 | pp16384   |       11797.73 |               11853.17 |        1.00 |
| RTX 3090   | llama 8B Q4_0   |                   8 | pp16384   |         423.95 |                 433.73 |        1.02 |
| RTX 3090   | llama 8B Q4_0   |                  16 | pp16384   |         854.51 |                 857.10 |        1.00 |
| RTX 3090   | llama 8B Q4_0   |                  32 | pp16384   |        1465.20 |                1458.56 |        1.00 |
| RTX 3090   | llama 8B Q4_0   |                  64 | pp16384   |        2247.54 |                2239.60 |        1.00 |
| RTX 3090   | llama 8B Q4_0   |                 128 | pp16384   |        2855.43 |                2845.70 |        1.00 |
| RTX 3090   | llama 8B Q4_0   |                 256 | pp16384   |        3298.58 |                3565.78 |        1.08 |
| RTX 3090   | llama 8B Q4_0   |                 512 | pp16384   |        3461.29 |                3694.70 |        1.07 |
| RTX 3090   | llama 8B Q4_0   |                1024 | pp16384   |        3501.73 |                3818.37 |        1.09 |
| RTX 3090   | llama 8B Q4_0   |                2048 | pp16384   |        3465.74 |                3816.77 |        1.10 |
| RTX 3090   | llama 8B Q4_0   |                4096 | pp16384   |        3477.07 |                3830.18 |        1.10 |
| RTX 3090   | llama 8B Q4_0   |                8192 | pp16384   |        3516.95 |                3835.93 |        1.09 |
| RTX 3090   | llama 8B Q4_0   |               16384 | pp16384   |        3413.82 |                3660.37 |        1.07 |
| RTX 3090   | phi2 3B F16     |                   8 | pp16384   |         535.35 |                 559.82 |        1.05 |
| RTX 3090   | phi2 3B F16     |                  16 | pp16384   |        1045.29 |                1069.07 |        1.02 |
| RTX 3090   | phi2 3B F16     |                  32 | pp16384   |        1944.48 |                1906.19 |        0.98 |
| RTX 3090   | phi2 3B F16     |                  64 | pp16384   |        3257.62 |                3172.20 |        0.97 |
| RTX 3090   | phi2 3B F16     |                 128 | pp16384   |        4727.21 |                4608.92 |        0.97 |
| RTX 3090   | phi2 3B F16     |                 256 | pp16384   |        5465.76 |                6035.44 |        1.10 |
| RTX 3090   | phi2 3B F16     |                 512 | pp16384   |        5834.88 |                6498.62 |        1.11 |
| RTX 3090   | phi2 3B F16     |                1024 | pp16384   |        5728.60 |                6542.47 |        1.14 |
| RTX 3090   | phi2 3B F16     |                2048 | pp16384   |        5710.91 |                6725.69 |        1.18 |
| RTX 3090   | phi2 3B F16     |                4096 | pp16384   |        5708.27 |                6719.02 |        1.18 |
| RTX 3090   | phi2 3B F16     |                8192 | pp16384   |        5712.93 |                6725.94 |        1.18 |
| RTX 3090   | phi2 3B F16     |               16384 | pp16384   |        5676.58 |                6731.76 |        1.19 |
| RTX 4090   | gemma 2B F16    |                   8 | pp16384   |        1204.39 |                1209.22 |        1.00 |
| RTX 4090   | gemma 2B F16    |                  16 | pp16384   |        2439.13 |                2168.01 |        0.89 |
| RTX 4090   | gemma 2B F16    |                  32 | pp16384   |        4686.86 |                3690.27 |        0.79 |
| RTX 4090   | gemma 2B F16    |                  64 | pp16384   |        7842.68 |                5194.76 |        0.66 |
| RTX 4090   | gemma 2B F16    |                 128 | pp16384   |       13437.09 |               11235.78 |        0.84 |
| RTX 4090   | gemma 2B F16    |                 256 | pp16384   |       20299.78 |               19070.07 |        0.94 |
| RTX 4090   | gemma 2B F16    |                 512 | pp16384   |       25635.41 |               25250.59 |        0.98 |
| RTX 4090   | gemma 2B F16    |                1024 | pp16384   |       24842.82 |               24671.27 |        0.99 |
| RTX 4090   | gemma 2B F16    |                2048 | pp16384   |       24328.35 |               24944.02 |        1.03 |
| RTX 4090   | gemma 2B F16    |                4096 | pp16384   |       24343.09 |               24838.79 |        1.02 |
| RTX 4090   | gemma 2B F16    |                8192 | pp16384   |       24244.10 |               24918.96 |        1.03 |
| RTX 4090   | gemma 2B F16    |               16384 | pp16384   |       24280.23 |               24897.67 |        1.03 |
| RTX 4090   | llama 8B Q4_0   |                   8 | pp16384   |         886.31 |                 904.86 |        1.02 |
| RTX 4090   | llama 8B Q4_0   |                  16 | pp16384   |        1590.18 |                1533.52 |        0.96 |
| RTX 4090   | llama 8B Q4_0   |                  32 | pp16384   |        2823.95 |                2694.92 |        0.95 |
| RTX 4090   | llama 8B Q4_0   |                  64 | pp16384   |        4963.35 |                4331.06 |        0.87 |
| RTX 4090   | llama 8B Q4_0   |                 128 | pp16384   |        6873.26 |                6517.14 |        0.95 |
| RTX 4090   | llama 8B Q4_0   |                 256 | pp16384   |        8614.69 |                8414.13 |        0.98 |
| RTX 4090   | llama 8B Q4_0   |                 512 | pp16384   |        9350.85 |                9609.05 |        1.03 |
| RTX 4090   | llama 8B Q4_0   |                1024 | pp16384   |        9363.54 |                9615.67 |        1.03 |
| RTX 4090   | llama 8B Q4_0   |                2048 | pp16384   |        8976.52 |                9201.50 |        1.03 |
| RTX 4090   | llama 8B Q4_0   |                4096 | pp16384   |        8971.77 |                9191.37 |        1.02 |
| RTX 4090   | llama 8B Q4_0   |                8192 | pp16384   |        8975.75 |                9185.91 |        1.02 |
| RTX 4090   | llama 8B Q4_0   |               16384 | pp16384   |        8971.71 |                9201.87 |        1.03 |
| RTX 4090   | phi2 3B F16     |                   8 | pp16384   |         690.30 |                 700.12 |        1.01 |
| RTX 4090   | phi2 3B F16     |                  16 | pp16384   |        1367.85 |                1315.42 |        0.96 |
| RTX 4090   | phi2 3B F16     |                  32 | pp16384   |        2677.82 |                2528.37 |        0.94 |
| RTX 4090   | phi2 3B F16     |                  64 | pp16384   |        5014.41 |                4357.40 |        0.87 |
| RTX 4090   | phi2 3B F16     |                 128 | pp16384   |        8427.73 |                7898.26 |        0.94 |
| RTX 4090   | phi2 3B F16     |                 256 | pp16384   |       12136.08 |               11752.74 |        0.97 |
| RTX 4090   | phi2 3B F16     |                 512 | pp16384   |       13649.55 |               14318.49 |        1.05 |
| RTX 4090   | phi2 3B F16     |                1024 | pp16384   |       15067.15 |               15920.39 |        1.06 |
| RTX 4090   | phi2 3B F16     |                2048 | pp16384   |       13654.12 |               14460.66 |        1.06 |
| RTX 4090   | phi2 3B F16     |                4096 | pp16384   |       13662.36 |               14443.62 |        1.06 |
| RTX 4090   | phi2 3B F16     |                8192 | pp16384   |       13664.72 |               14455.61 |        1.06 |
| RTX 4090   | phi2 3B F16     |               16384 | pp16384   |       13663.50 |               14460.25 |        1.06 |


</details>