@LostRuins 2025-02-17
Hi @JohannesGaessler thank you for this PR. I think it fixes a regression that has started since https://github.com/ggml-org/llama.cpp/pull/11583 was merged. Was just about to create a new issue/bug report here but you beat me to it, and I'll still note my findings for posterity even thought it (appears?) to have been fixed just in case. 

So basically since #11583 we have had [users](https://github.com/LostRuins/koboldcpp/issues/1372) complaining about incoherent outputs at high contexts with Mistral Small 3 based models, but surprisingly only on a **desktop RTX 4090** with flash attention enabled. I myself use a **laptop RTX 4090** which I thought had the exact same compute capability, but was never able to repro it locally, nor was anyone else running any older cards, we tested quite a few.

Finally I rented a desktop rtx 4090 on runpod, and I was able to reproduce the incoherent outputs issue.

I knew it was a problem with the new FA implementation because avoiding `ggml_cuda_flash_attn_ext_mma_f16` with
https://github.com/LostRuins/koboldcpp/commit/7502af9248291a4928d90426561ae775ddb57e84 and reverting to ` ggml_cuda_flash_attn_ext_wmma_f16(ctx, dst);` worked perfectly fine in all scenarios. 

I just couldn't understand why it was only limited to Desktop RTX 4090s, and was considering swapping back to the old FA wholesale. 

But it seems like your new PR solves it, at least from my initial tests. I'll update if I notice reports of this issue again.