@JohannesGaessler 2025-02-17
The reason the desktop 4090 specifically can produce incorrect results is the number of streaming multiprocessors. I changed the FlashAttention kernels to use stream-k decomposition which in most cases needs a fixup after the regular kernel because it can work on fractional tiles that may need to be combined. However, if the number of tiles can be exactly divided by the number of SMs then the fixup can be skipped. In my original PR I had swapped the arguments for the modulo operator so GPUs with an SM count that is a power of 2 could for some configurations skip the fixup incorrectly. And a desktop 4090 has 128 SMs, a laptop 4090 has 76 SMs.